7 changing the int to long

+++ Binomial.java
@@ -18,7 +18,7 @@ public class Binomial {
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 
 	    int[][] memo = new int[n + 1][k + 1];
 		for (int i = 0; i <= n; i++) {
@@ -29,7 +29,7 @@ public class Binomial {
     	return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, int[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
